requirements.txt
asgiref==3.2.10
boto3==1.16.13
botocore==1.19.13
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-storages==1.10.1
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
jmespath==0.10.0
Markdown==3.3.3
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11

twilio==6.51.0